requirements.txt
black==25.1.0
click==8.1.8
flake8==7.2.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pycodestyle==2.13.0
pyflakes==3.3.2
python-dotenv==1.1.0
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.13.0
